requirements.txt
Flask==2.2.5
Pillow==9.3.0
pdf2image==1.16.3
easyocr
numpy
easyocr
tqdm
